Template conversion (#2)
* Converted to using templates. There isn't a clean way to make a UE4 Plugin otherwise.  A typedef of WORLDSTATEPROPERTY_TYPE etc. is a bad idea because that makes the Plugin dependent on the main game code.

* make some BaseContext functions non-overrideable. These relate to the core state of a Context.

* - move headers into directories so that include paths must be specified from the root of the project directory. This keeps naming conflicts with other headers to a minimum by forcing a sort of "namespace"
- added some virtual destructors which UE4 complains about.

* - move pch back to root. It should not be included by consuming projects
- Moved STLTypes.h out of pch so that consumers are not forced to include pch.h

* remove all CPPs. IContext is a proper interface so there are no circular includes any more.
