+++ Fluid-HTNCPP.UnitTests/DomainTestContext.h
@@ -9,23 +9,23 @@ enum class DomainTestState
     HasB,
     HasC
 };
-class DomainTestWorldState : public IWorldState
+class DomainTestWorldState : public IWorldState<DomainTestState,uint8_t, DomainTestWorldState>
 {
     uint8_t MyState[3];
 
 public:
-    bool HasState(WORLDSTATEPROPERTY_ID_TYPE state, WORLDSTATEPROPERTY_VALUE_TYPE value) override
+    bool HasState(DomainTestState state, uint8_t value) 
     {
-        return (MyState[state] == value);
+        return (MyState[(int)state] == value);
     }
 
-    WORLDSTATEPROPERTY_VALUE_TYPE& GetState(WORLDSTATEPROPERTY_ID_TYPE state) override { return MyState[state]; }
+    uint8_t& GetState(DomainTestState state) { return MyState[(int)state]; }
 
-    void SetState(WORLDSTATEPROPERTY_ID_TYPE state, WORLDSTATEPROPERTY_VALUE_TYPE value) override { MyState[state] = value; }
+    void SetState(DomainTestState state, uint8_t value) { MyState[(int)state] = value; }
 
     int GetMaxPropertyCount() { return 3; }
 };
-class DomainTestContext : public BaseContext
+class DomainTestContext : public BaseContext<DomainTestState,uint8_t,DomainTestWorldState>
 {
     bool _done = false;
 
@@ -34,25 +34,26 @@ public:
 
     bool& Done() { return _done; }
 
-    bool HasStateOneParam(WORLDSTATEPROPERTY_ID_TYPE state)
+    bool HasStateOneParam(DomainTestState state)
     {
-        WORLDSTATEPROPERTY_VALUE_TYPE one = 1;
+        uint8_t one = 1;
         return BaseContext::HasState(state, one);
     }
-    bool HasStateOneParam(DomainTestState state) { return HasStateOneParam(static_cast<WORLDSTATEPROPERTY_ID_TYPE>(state)); }
 
     void SetStateDTS(DomainTestState state, int value)
     {
-        _WorldState->SetState(static_cast<WORLDSTATEPROPERTY_ID_TYPE>(state), static_cast<WORLDSTATEPROPERTY_VALUE_TYPE>(value));
+        _WorldState->SetState(static_cast<DomainTestState>(state), static_cast<uint8_t>(value));
     }
     void SetStateDTS(DomainTestState state, int value, bool dirty, EffectType eff)
     {
-        BaseContext::SetState(static_cast<WORLDSTATEPROPERTY_ID_TYPE>(state), static_cast<WORLDSTATEPROPERTY_VALUE_TYPE>(value),dirty,eff);
+        BaseContext::SetState(static_cast<DomainTestState>(state), static_cast<uint8_t>(value),dirty,eff);
     }
 
-    WORLDSTATEPROPERTY_VALUE_TYPE GetStateDTS(DomainTestState state) { return BaseContext::GetState(static_cast<WORLDSTATEPROPERTY_ID_TYPE>(state)); }
+    uint8_t GetStateDTS(DomainTestState state) { return BaseContext::GetState(static_cast<DomainTestState>(state)); }
 
 };
+typedef BaseContext<DomainTestState, uint8_t, DomainTestWorldState> BaseContextType;
+
 class MyDebugContext : public DomainTestContext
 {
 public:
