+++ Fluid-HTNCPP.UnitTests/FuncOperatorTests.cpp
@@ -2,18 +2,13 @@
 #include "CppUnitTest.h"
 #include "Contexts/BaseContext.h"
 #include "Operators/Operator.h"
+#include "DomainTestContext.h"
 
 using namespace Microsoft::VisualStudio::CppUnitTestFramework;
 
 using namespace FluidHTN;
 
 
-class TestContext : public BaseContext
-{
-	bool _Done = false;
-public:
-	bool& Done() { return _Done; }
-};
 
 namespace FluidHTNCPPUnitTests
 {
@@ -21,21 +16,21 @@ TEST_CLASS(FuncOperatorTests)
 {
     TEST_METHOD(UpdateDoesNothingWithoutFunctionPtr_ExpectedBehavior)
     {
-        TestContext ctx;
+        DomainTestContext ctx;
 		auto e = MakeSharedPtr<FuncOperator>(nullptr, nullptr);
 		e->Update(ctx);
 	} 
 
 TEST_METHOD(StopDoesNothingWithoutFunctionPtr_ExpectedBehavior)
 {
-    TestContext ctx;
+    DomainTestContext ctx;
     auto        e = MakeSharedPtr<FuncOperator>(nullptr, nullptr);
 
     e->Stop(ctx);
 }
 TEST_METHOD(UpdateReturnsStatusInternalFunctionPtr_ExpectedBehavior)
 {
-    TestContext ctx;
+    DomainTestContext ctx;
     auto        e = MakeSharedPtr<FuncOperator>([=](IContext&) { return TaskStatus::Success; }, nullptr);
 
     auto status = e->Update(ctx);
@@ -45,8 +40,8 @@ TEST_METHOD(UpdateReturnsStatusInternalFunctionPtr_ExpectedBehavior)
 
 TEST_METHOD(StopCallsInternalFunctionPtr_ExpectedBehavior)
 {
-    TestContext ctx;
-    auto e = MakeSharedPtr<FuncOperator>(nullptr, [](IContext&ctx ) { static_cast<TestContext&>(ctx).Done() = true; });
+    DomainTestContext ctx;
+    auto e = MakeSharedPtr<FuncOperator>(nullptr, [](IContext&ctx ) { static_cast<DomainTestContext&>(ctx).Done() = true; });
 
     e->Stop(ctx);
