+++ Fluid-HTNCPP.UnitTests/pch.h
@@ -24,7 +24,7 @@
 #include <cstdlib>
 #include <ctime>
 
-#include "STLTypes.h"
+#include "CoreIncludes/STLTypes.h"
 
 #ifndef FHTN_FATAL_EXCEPTION
 #define FHTN_FATAL_EXCEPTION(condition, msg)                                                                                          \
